worked on saving sleep info, loading/saving user info to the profile page. not complete yet, waiting on full implementation of profile page

+++ Fit.Natic/Database.cs
@@ -72,7 +72,7 @@ namespace Fit.Natic
         public int sleepTarget { get; set; }
         public int workoutTarget { get; set; }
         public int caloriesLogged { get; set; }
-        public int sleepLogged { get; set; }
+        public float sleepLogged { get; set; }
         public int workoutLogged { get; set; }
         public string notesLogged { get; set; }
